requirements.txt
langgraph>=0.2
langchain>=0.3
langchain-google-vertexai>=1.0
google-cloud-aiplatform>=1.60
gitpython>=3.1
javalang>=0.13
pygithub>=2.3
pydantic-settings>=2.5
loguru